requirements.txt
# Core Framework
Django==4.2.7

# Web Scraping
Scrapy==2.8.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# Asynchronous Task Queue
Celery==5.3.0
redis==4.5.5
django-celery-beat==2.5.0
django-celery-results==2.5.1

# API Development
djangorestframework==3.14.0
django-filter==23.1

# Caching
django-redis==5.3.0
